@@ -4,16 +4,17 @@ version: 0.3.0
 homepage: https://github.com/mattreid1/firebase_image
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  firebase_storage: ^7.0.0
-  sqflite: ^1.3.0
-  path: ^1.7.0
-  path_provider: ^1.6.24
+  firebase_core: ^1.0.0
+  firebase_storage: ^8.0.0
+  sqflite: ^2.0.0+2
+  path: ^1.8.0
+  path_provider: ^2.0.1
 
 dev_dependencies:
   flutter_test:
