@@ -17,13 +17,15 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({ Key? key, required this.title}) : super(key: key);
 
   final String title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
+
+
 
 class _MyHomePageState extends State<MyHomePage> {
   @override
@@ -42,7 +44,7 @@ class _MyHomePageState extends State<MyHomePage> {
           'gs://bucket123/userIcon123.jpg',
           shouldCache: true, // The image should be cached (default: True)
           maxSizeBytes: 3000 * 1000, // 3MB max file size (default: 2.5MB)
-          cacheRefreshStrategy: CacheRefreshStrategy.NEVER // Switch off update checking
+          cacheRefreshStrategy: CacheRefreshStrategy.NEVER,
         ),
         width: 100,
       ),
