@@ -23,7 +23,7 @@ class FirebaseImage extends ImageProvider<FirebaseImage> {
   final CacheRefreshStrategy cacheRefreshStrategy;
 
   /// Default: the default Firebase app. Specifies a custom Firebase app to make the request to the bucket from (optional)
-  final FirebaseApp firebaseApp;
+  FirebaseApp? firebaseApp = Firebase.app();
 
   /// The model for the image object
   final FirebaseImageObject _imageObject;
@@ -47,6 +47,7 @@ class FirebaseImage extends ImageProvider<FirebaseImage> {
           bucket: _getBucket(location),
           remotePath: _getImagePath(location),
           reference: _getImageRef(location, firebaseApp),
+          localPath: '',
         );
 
   /// Returns the image as bytes
@@ -64,21 +65,21 @@ class FirebaseImage extends ImageProvider<FirebaseImage> {
     return uri.path;
   }
 
-  static Reference _getImageRef(String location, FirebaseApp firebaseApp) {
+  static Reference _getImageRef(String location, FirebaseApp? firebaseApp) {
     FirebaseStorage storage = FirebaseStorage.instanceFor(
         app: firebaseApp, bucket: _getBucket(location));
     return storage.ref().child(_getImagePath(location));
   }
 
   Future<Uint8List> _fetchImage() async {
-    Uint8List bytes;
+    Uint8List? bytes;
     FirebaseImageCacheManager cacheManager = FirebaseImageCacheManager(
       cacheRefreshStrategy,
     );
 
     if (shouldCache) {
       await cacheManager.open();
-      FirebaseImageObject localObject =
+      FirebaseImageObject? localObject =
           await cacheManager.get(_imageObject.uri, this);
 
       if (localObject != null) {
@@ -96,11 +97,11 @@ class FirebaseImage extends ImageProvider<FirebaseImage> {
           await cacheManager.remoteFileBytes(_imageObject, this.maxSizeBytes);
     }
 
-    return bytes;
+    return bytes!;
   }
 
   Future<Codec> _fetchImageCodec() async {
-    return await PaintingBinding.instance
+    return await PaintingBinding.instance!
         .instantiateImageCodec(await _fetchImage());
   }
 
